#4 designing GameObject hierarchy

+++ src/main/java/com/jkojote/linear/engine/window/KeyCallback.java
@@ -2,5 +2,5 @@ package com.jkojote.linear.engine.window;
 
 public interface KeyCallback {
 
-    void perform(long key, int action, int mods);
+    void perform(int key, int action, int mods);
 }
